+++ lib/page_indexer.rb
